+++ FE/src/components/SettingContent/SaveModal/saveModal.scss
@@ -1,0 +1,65 @@
+@use '../../../styles/mixins.scss' as *;
+@use '../../../styles/values.scss' as *;
+
+.save__modal__overlay {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: rgba(0, 0, 0, 0.9);
+  z-index: 2;
+
+  .save__modal__content {
+    position: relative;
+    width: 50%;
+    height: 20%;
+    background-color: rgba(0, 0, 0, 0.9);
+    border: 1px solid #222;
+    border-radius: 6px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    .save__modal__title {
+      color: white;
+      font-size: 1.5em;
+    }
+    .save__ok__btn {
+      @include buttonDefaultSetting;
+      border: 1px solid $white;
+      color: $white;
+      position: absolute;
+      width: 5em;
+      height: 30px;
+      background-color: $modal-button-background;
+      right: 20%;
+      bottom: 1em;
+      font-size: 1em;
+
+      &:hover {
+        background-color: $light-black;
+        text-shadow: $selectedTextShadow;
+        box-shadow: $selectedBoxShadow;
+      }
+    }
+    .save__cancel__btn {
+      @include buttonDefaultSetting;
+      border: 1px solid $white;
+      color: $white;
+      position: absolute;
+      width: 5em;
+      height: 30px;
+      background-color: $modal-button-background;
+      right: 5%;
+      bottom: 1em;
+      font-size: 1em;
+
+      &:hover {
+        background-color: $light-black;
+        text-shadow: $selectedTextShadow;
+        box-shadow: $selectedBoxShadow;
+      }
+    }
+  }
+}
