+++ FE/src/pages/SettingPage/index.tsx
@@ -8,6 +8,7 @@ import CalendarSetting from '../../components/SettingContent/CalendarSetting';
 import CategorySetting from '../../components/SettingContent/CategorySetting';
 import CSVSetting from '../../components/SettingContent/CSVSetting';
 import useAccountBook from '../../service/useAccountBookSetting';
+import SaveModal from '../../components/SettingContent/SaveModal';
 
 import './settingPage.scss';
 
@@ -19,15 +20,6 @@ export default function SettingPage() {
 
   const history = useHistory();
   const accountBookId = history.location.state;
-
-  const onClickSaveModalOk = () => {
-    saveAction(updateData);
-    setSaveModal(false);
-  };
-
-  const onClickSaveModalCancel = () => {
-    setSaveModal(false);
-  };
 
   useAccountBook(accountBookId);
   return (
@@ -49,22 +41,11 @@ export default function SettingPage() {
       {settingType === 'category' && <CategorySetting />}
       {settingType === 'csv' && <CSVSetting />}
       {saveModal && (
-        <div className="save__modal__overlay">
-          <div className="save__modal__content">
-            <div className="save__modal__title">
-              변경사항을 저장 하시겠습니까?
-            </div>
-            <button className="save__ok__btn" onClick={onClickSaveModalOk}>
-              Ok
-            </button>
-            <button
-              className="save__cancel__btn"
-              onClick={onClickSaveModalCancel}
-            >
-              Cancel
-            </button>
-          </div>
-        </div>
+        <SaveModal
+          saveAction={saveAction}
+          updateData={updateData}
+          setSaveModal={setSaveModal}
+        />
       )}
     </div>
   );
